改了些插入bug

+++ src/main/java/com/ganbro/domain/entity/BookDetail.java
@@ -1,5 +1,6 @@
 package com.ganbro.domain.entity;
 
+import io.swagger.models.auth.In;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,12 +11,12 @@ import java.util.Date;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BookDetail {
-    private Long bookId;
+    private Integer bookId;
     private String bookName;
     private String publisher;
     private String publishDate;
     private Boolean isBorrowed;
-    private Long userId;
+    private Integer userId;
     // 数据库是Date型，这里用String简便代替
     private String startDate;
     private String dueDate;
